+++ JOSESwift/Sources/CryptoImplementation/EC.swift
@@ -32,6 +32,7 @@ internal enum ECError: Error {
     case encryptingFailed(description: String)
     case decryptingFailed(description: String)
     case invalidCurveDigestAlgorithm
+    case couldNotAllocateMemoryForSignature
 }
 
 /// Identifies the curve type parameter of a JWK representing an elliptic curve key
@@ -95,32 +96,15 @@ public enum ECCurveType: String, Codable {
 }
 
 fileprivate extension SignatureAlgorithm {
-    typealias DigestFunction = (
-            Optional<UnsafeRawPointer>,
-            UInt32,
-            Optional<UnsafeMutablePointer<UInt8>>
-    ) -> Optional<UnsafeMutablePointer<UInt8>>
 
-    func createDigest(input: Data) throws -> [UInt8] {
-        guard
-                let computedDigestLength = digestLength,
-                let computedDigestFunction = digestFunction
-                else {
-            throw ECError.invalidCurveDigestAlgorithm
-        }
-        var digest = [UInt8](repeating: 0, count: computedDigestLength)
-        _ = computedDigestFunction(Array(input), UInt32(input.count), &digest)
-        return digest
-    }
-
-    var digestLength: Int? {
+    var secKeyAlgorithm: SecKeyAlgorithm? {
         switch self {
         case .ES256:
-            return Int(CC_SHA256_DIGEST_LENGTH)
+            return .ecdsaSignatureMessageX962SHA256
         case .ES384:
-            return Int(CC_SHA384_DIGEST_LENGTH)
+            return .ecdsaSignatureMessageX962SHA384
         case .ES512:
-            return Int(CC_SHA512_DIGEST_LENGTH)
+            return .ecdsaSignatureMessageX962SHA512
         default:
             return nil
         }
@@ -134,19 +118,6 @@ fileprivate extension SignatureAlgorithm {
             return ECCurveType.P384
         case .ES512:
             return ECCurveType.P521
-        default:
-            return nil
-        }
-    }
-
-    var digestFunction: DigestFunction? {
-        switch self {
-        case .ES256:
-            return CC_SHA256
-        case .ES384:
-            return CC_SHA384
-        case .ES512:
-            return CC_SHA512
         default:
             return nil
         }
@@ -174,20 +145,34 @@ internal struct EC {
     /// - Returns: The signature.
     /// - Throws: `ECError` if any errors occur while signing the input data.
     static func sign(_ signingInput: Data, with privateKey: KeyType, and algorithm: SignatureAlgorithm) throws -> Data {
-        // Sign the input as raw elliptic curve coordinates using a hashing algorithm and a private key.
         guard let curveType = algorithm.curveType else {
             throw ECError.invalidCurveDigestAlgorithm
         }
-        let digest = try algorithm.createDigest(input: signingInput)
-        var signatureLength = curveType.signatureOctetLength
-        let signature = NSMutableData(length: signatureLength)!
-        let signatureBytes = signature.mutableBytes.assumingMemoryBound(to: UInt8.self)
-        let status = SecKeyRawSign(privateKey, .sigRaw, digest, digest.count, signatureBytes, &signatureLength)
-        if status != errSecSuccess {
-            throw ECError.signingFailed(description: "Error creating signature. (OSStatus: \(status))")
+        guard let secKeyAlgorithm = algorithm.secKeyAlgorithm else {
+            throw ECError.algorithmNotSupported
         }
 
-        return signature as Data
+        var cfErrorRef: Unmanaged<CFError>?
+        guard let signature = SecKeyCreateSignature(privateKey, secKeyAlgorithm, signingInput as CFData, &cfErrorRef) else {
+            if let error = cfErrorRef {
+                throw ECError.signingFailed(description: "Error creating signature. (CFError: \(error.takeRetainedValue()))")
+            }
+            fatalError("SecKeyCreateSignature returned nil but did not set CFError object.")
+        }
+
+        // unpack BER encoded ASN.1 format signature to raw format as specified for JWS
+        let ecSignatureTLV = [UInt8](signature as Data)
+        do {
+            let ecSignature = try ecSignatureTLV.read(.sequence)
+            let varlenR = try Data(ecSignature.read(.integer))
+            let varlenS = try Data(ecSignature.skip(.integer).read(.integer))
+            let fixlenR = Asn1IntegerConversion.toRaw(varlenR, of: curveType.coordinateOctetLength)
+            let fixlenS = Asn1IntegerConversion.toRaw(varlenS, of: curveType.coordinateOctetLength)
+
+            return fixlenR + fixlenS
+        } catch {
+            throw ECError.signingFailed(description: "Could not unpack ASN.1 EC signature.")
+        }
     }
 
     /// Verifies input data against a signature with a given elliptic curve algorithm and the corresponding public key.
@@ -200,18 +185,72 @@ internal struct EC {
     /// - Returns: True if the signature is verified, false if it is not verified.
     /// - Throws: `ECError` if any errors occur while verifying the input data against the signature.
     static func verify(_ verifyingInput: Data, against signature: Data, with publicKey: KeyType, and algorithm: SignatureAlgorithm) throws -> Bool {
-        // Verify the raw signature against an input with a hashing algorithm and public key.
+        // verify the raw signature against an input with a hashing algorithm and public key
         guard let curveType = algorithm.curveType else {
             throw ECError.invalidCurveDigestAlgorithm
         }
-        let digest = try algorithm.createDigest(input: verifyingInput)
-        let signatureBytes: [UInt8] = Array(signature)
-        let status = SecKeyRawVerify(publicKey, .sigRaw, digest, digest.count, signatureBytes, curveType.signatureOctetLength)
-        if status != errSecSuccess {
-            throw ECError.verifyingFailed(description: "Error validating signature. (OSStatus: \(status))")
+        guard let secKeyAlgorithm = algorithm.secKeyAlgorithm else {
+            throw ECError.algorithmNotSupported
+        }
+        if signature.count != (curveType.coordinateOctetLength * 2) {
+            throw ECError.verifyingFailed(description: "Signature is \(signature.count) bytes long instead of the expected \(curveType.coordinateOctetLength * 2).")
         }
 
-        return true
+        // pack raw signature as specified for JWS into BER encoded ASN.1 format
+        let fixlenR = signature.prefix(curveType.coordinateOctetLength)
+        let varlenR = [UInt8](Asn1IntegerConversion.fromRaw(fixlenR))
+        let fixlenS = signature.suffix(curveType.coordinateOctetLength)
+        let varlenS = [UInt8](Asn1IntegerConversion.fromRaw(fixlenS))
+        let asn1Signature = Data((varlenR.encode(as: .integer) + varlenS.encode(as: .integer)).encode(as: .sequence))
+
+        var cfErrorRef: Unmanaged<CFError>?
+        let isValid = SecKeyVerifySignature(publicKey, secKeyAlgorithm, verifyingInput as CFData, asn1Signature as CFData, &cfErrorRef)
+        if let error = cfErrorRef {
+            throw ECError.verifyingFailed(description: "Error verifying signature. (CFError: \(error.takeRetainedValue()))")
+        }
+        return isValid
+    }
+
+    // Converting integers to and from DER encoded ASN.1 as described here:
+    // https://docs.microsoft.com/en-us/windows/desktop/seccertenroll/about-integer
+    // This conversion is required because the Secure Enclave only supports generating ASN.1 encoded signatures,
+    // while the JWS Standard requires raw signatures, where the R and S are unsigned integers with a fixed length:
+    // https://github.com/airsidemobile/JOSESwift/pull/156#discussion_r292370209
+    // https://tools.ietf.org/html/rfc7515#appendix-A.3.1
+    internal struct Asn1IntegerConversion {
+        static func toRaw(_ data: Data, of fixedLength: Int) -> Data {
+            let varLength = data.count
+            if varLength > fixedLength + 1 {
+                fatalError("ASN.1 integer is \(varLength) bytes long when it should be < \(fixedLength + 1).")
+            }
+            if varLength == fixedLength + 1 {
+                assert(data.first == 0)
+                return data.dropFirst()
+            }
+            if varLength == fixedLength {
+                return data
+            }
+            if varLength < fixedLength {
+                // pad to fixed length using 0x00 bytes
+                return Data(count: fixedLength - varLength) + data
+            }
+            fatalError("Unable to parse ASN.1 integer. This should be unreachable.")
+        }
+
+        static func fromRaw(_ data: Data) -> Data {
+            assert(data.count > 0)
+            let msb: UInt8 = 0b1000_0000
+            // drop all leading zero bytes
+            let varlen = data.drop { $0 == 0}
+            guard let firstNonZero = varlen.first else {
+                // all bytes were zero so the encoded value is zero
+                return Data(count: 1)
+            }
+            if (firstNonZero & msb) == msb {
+                return Data(count: 1) + varlen
+            }
+            return varlen
+        }
     }
 
 }