+++ JOSESwift/Sources/CryptoImplementation/HMAC.swift
@@ -24,6 +24,10 @@
 import Foundation
 import CommonCrypto
 
+enum HMACError: Error {
+    case inputMustBeGreaterThanZero
+}
+
 fileprivate extension HMACAlgorithm {
     var ccAlgorithm: CCAlgorithm {
         switch self {
@@ -43,19 +47,32 @@ internal struct HMAC {
     ///
     /// - Parameters:
     ///   - input: The input to calculate a HMAC for.
-    ///   - key: The key used in the HMAC algorithm.
+    ///   - key: The key used in the HMAC algorithm. Must not be empty.
     ///   - algorithm: The algorithm used to calculate the HMAC.
     /// - Returns: The calculated HMAC.
-    static func calculate(from input: Data, with key: Data, using algorithm: HMACAlgorithm) -> Data {
+    static func calculate(from input: Data, with key: Data, using algorithm: HMACAlgorithm) throws -> Data {
+        guard input.count > 0 else {
+            throw HMACError.inputMustBeGreaterThanZero
+        }
+
         var hmacOutData = Data(count: algorithm.outputLength)
 
+        // Force unwrapping is ok, since input count is checked and key and algorithm are assumed not to be empty.
+        // From the docs: If the baseAddress of this buffer is nil, the count is zero.
+        // swiftlint:disable force_unwrapping
         hmacOutData.withUnsafeMutableBytes { hmacOutBytes in
             key.withUnsafeBytes { keyBytes in
                 input.withUnsafeBytes { inputBytes in
-                    CCHmac(algorithm.ccAlgorithm, keyBytes, key.count, inputBytes, input.count, hmacOutBytes)
+                    CCHmac(
+                        algorithm.ccAlgorithm,
+                        keyBytes.baseAddress!, key.count,
+                        inputBytes.baseAddress!, input.count,
+                        hmacOutBytes.baseAddress!
+                    )
                 }
             }
         }
+        // swiftlint:enable force_unwrapping
 
         return hmacOutData
     }