+++ Tests/DataRSAPublicKeyTests.swift
@@ -1,3 +1,4 @@
+// swiftlint:disable force_unwrapping
 //
 //  DataRSAPublicKeyTests.swift
 //  Tests
@@ -90,5 +91,5 @@ class DataRSAPublicKeyTests: RSACryptoTestCase {
 
         XCTAssertEqual(data, expectedData)
     }
-    
+
 }