+++ Tests/JWEHeaderTests.swift
@@ -1,3 +1,4 @@
+// swiftlint:disable force_unwrapping
 //
 //  JWEHeaderTests.swift
 //  Tests
@@ -50,7 +51,7 @@ class JWEHeaderTests: XCTestCase {
 
         XCTAssertEqual(header.parameters["enc"] as? String, SymmetricContentAlgorithm.A256CBCHS512.rawValue)
         XCTAssertEqual(header.parameters["alg"] as? String, AsymmetricKeyAlgorithm.RSA1_5.rawValue)
-        XCTAssertEqual(header.data(), try! JSONSerialization.data(withJSONObject: parameterDictRSA, options: []))
+        XCTAssertEqual(header.data().count, try! JSONSerialization.data(withJSONObject: parameterDictRSA, options: []).count)
     }
 
     func testInitRSAOAEPWithParameters() {
@@ -58,7 +59,7 @@ class JWEHeaderTests: XCTestCase {
 
         XCTAssertEqual(header.parameters["enc"] as? String, SymmetricContentAlgorithm.A256CBCHS512.rawValue)
         XCTAssertEqual(header.parameters["alg"] as? String, AsymmetricKeyAlgorithm.RSAOAEP.rawValue)
-        XCTAssertEqual(header.data(), try! JSONSerialization.data(withJSONObject: parameterDictRSAOAEP, options: []))
+        XCTAssertEqual(header.data().count, try! JSONSerialization.data(withJSONObject: parameterDictRSAOAEP, options: []).count)
     }
 
     func testInitRSAOAEP256WithParameters() {
@@ -66,7 +67,7 @@ class JWEHeaderTests: XCTestCase {
         
         XCTAssertEqual(header.parameters["enc"] as? String, SymmetricContentAlgorithm.A256CBCHS512.rawValue)
         XCTAssertEqual(header.parameters["alg"] as? String, AsymmetricKeyAlgorithm.RSAOAEP256.rawValue)
-        XCTAssertEqual(header.data(), try! JSONSerialization.data(withJSONObject: parameterDictRSAOAEP256, options: []))
+        XCTAssertEqual(header.data().count, try! JSONSerialization.data(withJSONObject: parameterDictRSAOAEP256, options: []).count)
     }
 
     func testInitRSA1WithData() {
@@ -101,7 +102,7 @@ class JWEHeaderTests: XCTestCase {
 
         XCTAssertEqual(header.parameters["enc"] as? String, SymmetricContentAlgorithm.A256CBCHS512.rawValue)
         XCTAssertEqual(header.parameters["alg"] as? String, AsymmetricKeyAlgorithm.direct.rawValue)
-        XCTAssertEqual(header.data(), try! JSONSerialization.data(withJSONObject: parameterDictDirect, options: []))
+        XCTAssertEqual(header.data().count, try! JSONSerialization.data(withJSONObject: parameterDictDirect, options: []).count)
     }
 
     func testInitDirectWithData() {
@@ -116,7 +117,7 @@ class JWEHeaderTests: XCTestCase {
     func testInitWithAlgAndEncRSA1() {
         let header = JWEHeader(algorithm: .RSA1_5, encryptionAlgorithm: .A256CBCHS512)
 
-        XCTAssertEqual(header.data(), try! JSONSerialization.data(withJSONObject: parameterDictRSA, options: []))
+        XCTAssertEqual(header.data().count, try! JSONSerialization.data(withJSONObject: parameterDictRSA, options: []).count)
         XCTAssertEqual(header.parameters["alg"] as? String, AsymmetricKeyAlgorithm.RSA1_5.rawValue)
         XCTAssertEqual(header.parameters["enc"] as? String, SymmetricContentAlgorithm.A256CBCHS512.rawValue)
 
@@ -129,7 +130,7 @@ class JWEHeaderTests: XCTestCase {
     func testInitWithAlgAndEncRSAOAEP() {
         let header = JWEHeader(algorithm: .RSAOAEP, encryptionAlgorithm: .A256CBCHS512)
 
-        XCTAssertEqual(header.data(), try! JSONSerialization.data(withJSONObject: parameterDictRSAOAEP, options: []))
+        XCTAssertEqual(header.data().count, try! JSONSerialization.data(withJSONObject: parameterDictRSAOAEP, options: []).count)
         XCTAssertEqual(header.parameters["alg"] as? String, AsymmetricKeyAlgorithm.RSAOAEP.rawValue)
         XCTAssertEqual(header.parameters["enc"] as? String, SymmetricContentAlgorithm.A256CBCHS512.rawValue)
 
@@ -141,8 +142,8 @@ class JWEHeaderTests: XCTestCase {
 
     func testInitWithAlgAndEncRSAOAEP256() {
         let header = JWEHeader(algorithm: .RSAOAEP256, encryptionAlgorithm: .A256CBCHS512)
-        
-        XCTAssertEqual(header.data(), try! JSONSerialization.data(withJSONObject: parameterDictRSAOAEP256, options: []))
+
+        XCTAssertEqual(header.data().count, try! JSONSerialization.data(withJSONObject: parameterDictRSAOAEP256, options: []).count)
         XCTAssertEqual(header.parameters["alg"] as? String, AsymmetricKeyAlgorithm.RSAOAEP256.rawValue)
         XCTAssertEqual(header.parameters["enc"] as? String, SymmetricContentAlgorithm.A256CBCHS512.rawValue)
         
@@ -230,7 +231,7 @@ class JWEHeaderTests: XCTestCase {
         header.crit = crit
         header.zip = zip
 
-        XCTAssertEqual(header.data(), try! JSONSerialization.data(withJSONObject: header.parameters, options: []))
+        XCTAssertEqual(header.data().count, try! JSONSerialization.data(withJSONObject: header.parameters, options: []).count)
 
         XCTAssertEqual(header.parameters["jku"] as? URL, jku)
         XCTAssertEqual(header.jku, jku)