+++ Tests/RSAPublicKeyToDataTests.swift
@@ -25,7 +25,7 @@ import XCTest
 @testable import JOSESwift
 
 class RSAPublicKeyToDataTests: RSACryptoTestCase {
-    
+
     func testpublicKeyAlice2048ToData() {
         let jwk = RSAPublicKey(modulus: expectedModulus2048Base64, exponent: expectedExponentBase64)
         let data = try! jwk.converted(to: Data.self)
@@ -39,5 +39,5 @@ class RSAPublicKeyToDataTests: RSACryptoTestCase {
 
         XCTAssertEqual(data, publicKey4096Data)
     }
-    
+
 }