+++ Tests/SecKeyECPublicKeyTests.swift
@@ -1,3 +1,4 @@
+// swiftlint:disable force_unwrapping
 //
 //  SecKeyECPublicKeyTests.swift
 //  Tests
@@ -53,13 +54,10 @@ class SecKeyECPublicKeyTests: ECCryptoTestCase {
         }
     }
 
-    func testPublicKeyFromPublicComponents() {
-        allTestData.forEach { testData in
+    func testPublicKeyFromPublicComponents() throws {
+        try allTestData.forEach { testData in
             let components = (testData.expectedCurveType, testData.expectedXCoordinate, testData.expectedYCoordinate)
-            guard let secKey = try? SecKey.representing(ecPublicKeyComponents: components) else {
-                XCTFail()
-                return
-            }
+            let secKey = try SecKey.representing(ecPublicKeyComponents: components)
 
             let data = SecKeyCopyExternalRepresentation(secKey, nil)! as Data
             let dataExpected = SecKeyCopyExternalRepresentation(testData.publicKey, nil)! as Data